making neovim faster again
The horizontal and vertical scrolling was very laggy, especially in Vue projects. This was due to the fact that Volar requires `capabilities.workspace.didChangeWatchedFiles.dynamicRegistration` set to true, and the current implementation of _watchfiles was causing that lag/stutter. The horizontal scrolling lag/stutter is caused by long lines and having nowrap activated. When performing horizontal scrolling, seems like Neovim recalculates with regex the highlighting (I'm not sure about this). However, setting wrap resolves the issue.